manage all zones
