+++ playbook-tests/test-connect.yml
@@ -1,0 +1,11 @@
+---
+- name: Test connection to Scaleway hosts
+  hosts: all
+  gather_facts: no
+  tasks:
+    - name: Ping host
+      ping:
+
+    - name: Show host info
+      debug:
+        msg: "Connected to {{ inventory_hostname }} ({{ ansible_host }})"
